Minor fixes to get rid of warnings from BCB.

+++ TAO/TAO_IDL/be/be_visitor_array/array.cpp
@@ -110,7 +110,7 @@ be_visitor_array::visit_string (be_string *node)
 {
   TAO_OutStream *os = this->ctx_->stream ();
 
-  if (node->width () == sizeof (char))
+  if (node->width () == (long) sizeof (char))
     {
       *os << "TAO_String_Manager";
     }
